Big input system refactor
- Update from SDL 1.2 to SDL 2.
- Move input handling to p8_input.c.
- Added basic keyboard and mouse event 'queue'.
