--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "SDL-1.2"]
-	path = SDL-1.2
-	url = https://github.com/libsdl-org/SDL-1.2
