+++ Makefile.linux
@@ -1,10 +1,9 @@
-# SDL 1.2 (Option A): use system-installed SDL dev package via pkg-config (preferred)
-# Fallback to sdl-config if pkg-config is unavailable.
-SDL_CFLAGS := $(shell pkg-config --cflags sdl 2>/dev/null || sdl-config --cflags 2>/dev/null)
-SDL_LIBS   := $(shell pkg-config --libs   sdl 2>/dev/null || sdl-config --libs   2>/dev/null)
+# SDL2: use system-installed SDL2 dev package via pkg-config.
+SDL_CFLAGS := $(shell pkg-config --cflags sdl2 2>/dev/null)
+SDL_LIBS   := $(shell pkg-config --libs   sdl2 2>/dev/null)
 
 ifeq ($(strip $(SDL_CFLAGS)),)
-  $(error SDL development files not found. Install libsdl1.2-dev and pkg-config (or ensure sdl-config is available).)
+  $(error SDL2 development files not found. Install libsdl2-dev and pkg-config.)
 endif
 
 DEFINES += -DLUA_USE_POSIX
