+++ Makefile.mac
@@ -1,10 +1,9 @@
-# SDL 1.2 (Option A): use system-installed SDL dev package via pkg-config (preferred)
-# Fallback to sdl-config if pkg-config is unavailable.
-SDL_CFLAGS := $(shell pkg-config --cflags sdl 2>/dev/null || sdl-config --cflags 2>/dev/null)
-SDL_LIBS   := $(shell pkg-config --libs   sdl 2>/dev/null || sdl-config --libs   2>/dev/null)
+# SDL2: use system-installed SDL2 dev package via pkg-config.
+SDL_CFLAGS := $(shell pkg-config --cflags sdl2 2>/dev/null)
+SDL_LIBS   := $(shell pkg-config --libs   sdl2 2>/dev/null)
 
 ifeq ($(strip $(SDL_CFLAGS)),)
-  $(error SDL development files not found. On GitHub macOS runners: brew install pkg-config sdl12-compat)
+  $(error SDL2 development files not found. On GitHub macOS runners: brew install pkg-config sdl2)
 endif
 
 DEFINES += -DLUA_USE_POSIX -DLUA_USE_DLOPEN -Dmain=SDL_main
