+++ Makefile.windows
@@ -1,10 +1,9 @@
-# SDL 1.2: use system-installed SDL dev package via pkg-config (preferred)
-# Fallback to sdl-config if pkg-config is unavailable.
-SDL_CFLAGS_RAW := $(shell pkg-config --cflags sdl 2>/dev/null || sdl-config --cflags 2>/dev/null)
-SDL_LIBS_RAW   := $(shell pkg-config --libs   sdl 2>/dev/null || sdl-config --libs   2>/dev/null)
+# SDL2: use system-installed SDL2 dev package via pkg-config.
+SDL_CFLAGS_RAW := $(shell pkg-config --cflags sdl2 2>/dev/null)
+SDL_LIBS_RAW   := $(shell pkg-config --libs   sdl2 2>/dev/null)
 
 ifeq ($(strip $(SDL_CFLAGS_RAW)),)
-  $(error SDL development files not found. In MSYS2 UCRT64 install mingw-w64-ucrt-x86_64-SDL and mingw-w64-ucrt-x86_64-pkgconf.)
+  $(error SDL development files not found. In MSYS2 UCRT64 install mingw-w64-ucrt-x86_64-SDL2 and mingw-w64-ucrt-x86_64-pkgconf.)
 endif
 
 DEFINES += -DLUA_USE_WINDOWS -D_CRT_SECURE_NO_WARNINGS -DNO_STDIO_REDIRECT
