--- SDL-1.2
@@ -1,1 +1,0 @@
-SDL-1.2-linux
